extract connection assertions
